pyproject.toml
[tool.poetry]
name = "inclusion&exclusionfilter"
version = "0.1.0"
description = ""
authors = ["Joel-KSanthosh <joelksanthosh02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
